Add cast to int

+++ Snowflake.Data/Core/CredentialManager/Infrastructure/SFCredentialManagerWindowsNativeImpl.cs
@@ -46,7 +46,7 @@ namespace Snowflake.Data.Core.CredentialManager.Infrastructure
             using (var critCred = new CriticalCredentialHandle(nCredPtr))
             {
                 var cred = critCred.GetCredential();
-                if (cred.CredentialBlob.Length > cred.CredentialBlobSize / 2)
+                if (cred.CredentialBlob.Length > (int)cred.CredentialBlobSize / 2)
                 {
                     return cred.CredentialBlob.Substring(0, (int)(cred.CredentialBlobSize / 2));
                 }
